Make make_gif and dump_frames compatible.
Because of the way that scitools makes gifs, the frames
are ordered based on their filenames on disk. Their order
is determined based on a pattern syntax and alphabetical
ordering. To make it use numerical ordering, all frame
filenames have to have the same padding. Ie

frame_0001.jpg
frame_0002.jpg

If this is not the case, then

frame11.jpg

will come before

frame 6.jpg

The caveat is that I fixed the width to be 4. So, if we ever
want to make animated gifs with more than 9999 frames, we'll
have to adjust.